+++ CHANGELOG.md
@@ -2,3 +2,8 @@
 All notable changes to this project will be documented in this file, which follows the guidelines
 on [Keep a CHANGELOG](http://keepachangelog.com/). This project adheres to
 [Semantic Versioning](http://semver.org/).
+
+## [Unreleased]
+
+### Added
+- Replay Event Stream tool